requirements.txt
Django==3.2.16
django-auth-ldap
django-currentuser
django-datatables-view
django-guardian
django-jsonview
django-leaflet
django-user-agents
djangorestframework
piccolo-theme
Pillow
psycopg2-binary
PyYAML
requests
Sphinx
ua-parser
user-agents
